[CLN] reorder data classes
