+++ test/test_input_data/test_data_classes.py
@@ -8,13 +8,15 @@ import pandas as pn
 import matplotlib.pyplot as plt
 import pytest
 
-import gempy_lite.core.data_modules.geometric_data
-import gempy_lite.core.data_modules.stack
+import gempy_lite.core.kernel_data
+import gempy_lite.core.kernel_data.geometric_data
+import gempy_lite.core.kernel_data.stack
+import gempy_lite.core.model_data
 
 
 @pytest.fixture(scope='module')
 def create_faults():
-    faults = gempy_lite.core.data_modules.stack.Faults()
+    faults = gempy_lite.core.kernel_data.stack.Faults()
     return faults
 
 
@@ -22,7 +24,7 @@ def create_faults():
 def create_series(create_faults):
     faults = create_faults
 
-    series = gempy_lite.core.data_modules.stack.Series(faults)
+    series = gempy_lite.core.kernel_data.stack.Series(faults)
     series.set_series_index(['foo', 'foo2', 'foo5', 'foo7'])
     series.add_series('foo3')
     series.delete_series('foo2')
@@ -43,7 +45,7 @@ def create_series(create_faults):
 @pytest.fixture(scope='module')
 def create_surfaces(create_series):
     series = create_series
-    surfaces = gp.Surfaces(series)
+    surfaces = gempy_lite.core.kernel_data.Surfaces(series)
     surfaces.set_surfaces_names(['foo', 'foo2', 'foo5'])
 
     print(series)
@@ -140,7 +142,7 @@ def create_surface_points(create_surfaces, create_series):
     # These two DataFrames (df from now on) will contain the individual information of each point at an interface or
     # orientation. Some properties of this table are mapped from the *df* below.
     surfaces = create_surfaces
-    surface_points = gempy_lite.core.data_modules.geometric_data.SurfacePoints(surfaces)
+    surface_points = gempy_lite.core.kernel_data.geometric_data.SurfacePoints(surfaces)
 
     print(surface_points)
 
@@ -169,7 +171,7 @@ def create_orientations(create_surfaces, create_series):
     surfaces = create_surfaces
 
     # ### Orientations
-    orientations = gempy_lite.core.data_modules.geometric_data.Orientations(surfaces)
+    orientations = gempy_lite.core.kernel_data.geometric_data.Orientations(surfaces)
 
     print(orientations)
 
@@ -204,7 +206,7 @@ def create_orientations(create_surfaces, create_series):
 
 
 def test_add_orientation_with_pole(create_surfaces):
-    orientations = gempy_lite.core.data_modules.geometric_data.Orientations(create_surfaces)
+    orientations = gempy_lite.core.kernel_data.geometric_data.Orientations(create_surfaces)
     orientations.add_orientation(1, 1, 1, 'foo', pole_vector=(1, 0, 1))
     orientations.add_orientation(2, 2, 2, 'foo', orientation=(0, 0, 1))
     orientations.add_orientation(1, 1, 1, 'foo', pole_vector=(.45, 0, .45))
@@ -224,7 +226,7 @@ def create_grid():
 
 @pytest.fixture(scope='module')
 def create_rescaling(create_surface_points, create_orientations, create_grid):
-    rescaling = gempy_lite.core.data_modules.geometric_data.RescaledData(create_surface_points, create_orientations, create_grid)
+    rescaling = gempy_lite.core.model_data.RescaledData(create_surface_points, create_orientations, create_grid)
     return rescaling
 
 
@@ -232,8 +234,8 @@ def create_rescaling(create_surface_points, create_orientations, create_grid):
 def create_additional_data(create_surface_points, create_orientations, create_grid, create_faults,
                            create_surfaces, create_rescaling):
 
-    ad = gp.AdditionalData(create_surface_points, create_orientations, create_grid, create_faults,
-                           create_surfaces, create_rescaling)
+    ad = gempy_lite.core.model_data.AdditionalData(create_surface_points, create_orientations, create_grid, create_faults,
+                                                   create_surfaces, create_rescaling)
     return ad
 
 
@@ -259,7 +261,7 @@ class TestDataManipulation:
 
 
 def test_stack():
-    stack = gempy_lite.core.data_modules.stack.Stack()
+    stack = gempy_lite.core.kernel_data.stack.Stack()
     stack.set_series_index(['foo', 'foo2', 'foo5', 'foo7'])
     stack.add_series('foo3')
     stack.delete_series('foo2')
